requirements.txt
flask
SQLAlchemy
python-dateutil
pandas
seaborn
scipy
RPi-GPIO-I2C-LCD
pyre-check
python-dotenv
pid